% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hirlite.R
\name{hirlite}
\alias{hirlite}
\title{Interface to rlite}
\usage{
hirlite(...)
}
\arguments{
\item{...}{Named configuration options passed to
  \code{\link{redis_config}}, used to create the environment
  (notable keys include \code{host}, \code{port}, and the
  environment variable \code{REDIS_URL})}
}
\description{
Create an interface to rlite, with a generated interface to all
rlite commands (using \code{RedisAPI}).
}
\examples{
# Only run if a Redis server is running
if (redis_available()) {
  r <- hirlite()
  r$PING()
  r$SET("foo", "bar")
  r$GET("foo")
}
}

