% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hiredis.R
\name{parse_redis_info}
\alias{parse_redis_info}
\title{Parse Redis INFO}
\usage{
parse_redis_info(x)
}
\arguments{
\item{x}{character string}
}
\description{
Parse information returned by \code{INFO}
}

