% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcpp_redis.R
\name{rcpp_redis}
\alias{rcpp_redis}
\title{Interface to RcppRedis}
\usage{
rcpp_redis(host = "127.0.0.1", port = 6379)
}
\arguments{
\item{host}{Hostname (default is the localhost)}

\item{port}{Port to connect on}
}
\description{
Interface to RcppRedis.  This is the drop in replacement for
\code{\link{rlite_context}}
}

