% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdb.R
\name{rdb}
\alias{rdb}
\title{Create an rdb object}
\usage{
rdb(path = ":memory:", hiredis = NULL)
}
\arguments{
\item{path}{to a persistent object or the magic string
\code{":memory:"} for an in-memory database.}

\item{hiredis}{Alternative hiredis object implementation}
}
\description{
Create an rdb object
}
\examples{
r <- rdb()
r$set("foo", runif(10))
r$get("foo")
r$keys()
r$del("foo")
r$keys()
}

