% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/help.R
\name{redis_commands}
\alias{redis_commands}
\alias{redis_commands_groups}
\title{List of valid redis commands}
\usage{
redis_commands(group = NULL)

redis_commands_groups()
}
\arguments{
\item{group}{Optional group name (run \code{redis_commands_groups}
for valid names here).}
}
\description{
List of valid redis commands supported by rlite/rrlite; note that
this is a \emph{subset} of all redis commands.
}
\examples{
redis_commands()
redis_commands_groups()
redis_commands("string")
}

