% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/help.R
\name{redis_help}
\alias{redis_help}
\title{Print help for redis command}
\usage{
redis_help(command)
}
\arguments{
\item{command}{Name of a command (see \code{\link{redis_commands}})}
}
\description{
Print help for redis command
}
\examples{
redis_help("SET")
}

