% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlite.R
\name{rlite}
\alias{rlite}
\title{Create an rlite object}
\usage{
rlite(path = ":memory:")
}
\arguments{
\item{path}{Path to a persistent object or the magic string
\code{":memory:"} for an in-memory database.}
}
\description{
Create an rlite object
}
\examples{
r <- rlite()
r$SET("foo", "bar")
r$GET("foo")
}

