% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlite_config.R
\name{rlite_config}
\alias{rlite_config}
\title{rlite configuration}
\usage{
rlite_config(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{redis_config}}; see
that file for more information.}
}
\description{
rlite configuration settings.  Based on the \code{redis_config}
function but with additional tweaks for rlite.  The differences
between this configuration and \code{\link{redis_config}} is that:
}
\details{
\itemize{

\item{\code{RLITE_URL} takes precendence over \code{REDIS_URL} if
both are present (otherwise \code{REDIS_URL} will still be used).}

\item{A host of \code{localhost} or \code{127.0.0.1}, which is
\code{redis_config}'s default, will map to a filename of
\code{:memory:} for a transient in-memory store.}
}

The \code{port} entry will be ignored, but the \code{password} and
\code{db} entries will be used if present.  \code{path} is
equivalent to \code{host}.
}

