% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlite_context.R
\name{rlite_context}
\alias{rlite_context}
\title{Create an rlite_context object}
\usage{
rlite_context(path = ":memory:")
}
\arguments{
\item{path}{Path to a persistent object or the magic string
\code{":memory:"} for an in-memory database.}
}
\description{
Create an \code{rlite_context} object
}
\examples{
r <- rlite_context()
r$run(c("SET", "foo", "bar"))
r$run(c("GET", "foo"))
}

