% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objects.R
\name{to_redis}
\alias{from_redis}
\alias{to_redis}
\title{Read/write R objects to Redis}
\usage{
to_redis(object, key, db, ...)

from_redis(key, db, ...)
}
\arguments{
\item{object}{An R object (must be a data.frame at present)}

\item{key}{The key to store it in}

\item{db}{The database to store it in}

\item{...}{Additional arguments to methods}
}
\description{
Support for writing and reading R objects (currently data.frames)
to a redis instance.
}

