% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hirlite.R
\name{hirlite}
\alias{hirlite}
\title{Create an hirlite object}
\usage{
hirlite(path = ":memory:")
}
\arguments{
\item{path}{Path to a persistent object or the magic string
\code{":memory:"} for an in-memory database.}
}
\description{
Create an hirlite object
}
\examples{
r <- hirlite()
r$SET("foo", "bar")
r$GET("foo")
}

