% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hirlite.R
\name{hirlite}
\alias{hirlite}
\alias{rlite_available}
\title{Interface to rlite}
\usage{
hirlite(...)

rlite_available(...)
}
\arguments{
\item{...}{Named configuration options passed to
  \code{\link{redis_config}}, used to create the environment
  (notable keys include \code{host}, \code{port}, and the
  environment variable \code{REDIS_URL}).  In addition to the
  \code{RedisAPI} treatment of the configuration, \code{RLITE_URL}
  takes precendence over \code{REDIS_URL}, and a host of
  \code{localhost} or \code{127.0.0.1} will be treated as an
  in-memory database (\code{:memory:}).}
}
\description{
Create an interface to rlite, with a generated interface to all
rlite commands (using \code{RedisAPI}).
}
\examples{
r <- hirlite()
r$PING()
r$SET("foo", "bar")
r$GET("foo")
}

