% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdb.R
\name{rdb}
\alias{rdb}
\title{Create an rlite object}
\usage{
rdb(path = ":memory:")
}
\arguments{
\item{path}{to a persistent object or the magic string
\code{":memory:"} for an in-memory database.}
}
\description{
Create an rlite object
}
\examples{
r <- rdb()
r$set("foo", runif(10))
r$get("foo")
r$list()
r$del("foo")
r$list()
}

